+++ BOCCHI/Modules/CriticalEncounters/CriticalEncountersModule.cs
@@ -58,7 +58,7 @@ public class CriticalEncountersModule : Module
         return true;
     }
 
-    public override void OnTerritoryChanged(ushort id)
+    public override void OnTerritoryChanged(uint id)
     {
         CriticalEncounters.Clear();
     }
