+++ BOCCHI/Modules/Currency/CurrencyModule.cs
@@ -30,7 +30,7 @@ public class CurrencyModule(Plugin plugin, Config config) : Module(plugin, confi
         Tracker.Tick(context.Framework);
     }
 
-    public override void OnTerritoryChanged(ushort _)
+    public override void OnTerritoryChanged(uint _)
     {
         Tracker.Reset();
     }
