+++ BOCCHI/Modules/Debug/Panels/Panel.cs
@@ -12,7 +12,7 @@ public abstract class Panel
     {
     }
 
-    public virtual void OnTerritoryChanged(ushort id, DebugModule module)
+    public virtual void OnTerritoryChanged(uint id, DebugModule module)
     {
     }
 }
